[FIX] stock: access error on stock.quant
Usecase to reproduce:
- Connect on demo user (inventory user)
- Go on product page and click on stat button for 'quantity on hand'

> AccessError on decimal.precision object

due to commit 414c6cf924600eb1c54ea0332f0a026ea5ec5107
When the user click on 'report inventory' button it will garbage
collect the remaining quants without quantity. In order to do it,
it needs the precision rounding in order to have a rounding and do
not let quant without significant quantity.

closes odoo/odoo#34780

Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>