[REF] stock_landed_costs: default journal id
Set a default journal on the company, use it on the landed costs form.
I was not able to directly set the value on the filed with a post init
hook because the stock journal is generated a while after the landed
costs installation in the l10n_generic_coa module.

task-59011