[IMP] delivery, sale: improvements in recomputing of shipping cost
The "Recompute shipping cost" button will now reopen the wizard
to be able to change carrier.

When adding to the order, it will replace the current shipping cost.

Relabelled the "Recompute shipping cost" to “Update Shipping Cost”
and moved it to the right.

Shipping method field will have radio widget and cost will be editable
in the wizard.

Also, it was not allowed to delete a delivery line when the order was
confirmed ('sale' state). So here, as we want to allow user to change the
carrier even after the order is confirmed, we have allowed delivery line
to be deleted so that new line is added in place of that.

task-2033229