[FIX] project_timesheet_holidays: Allow to create a company
It is currently impossible to create a company.
The reason being that the newly created company is
not yet in context `allowed_company_ids`.

Therefore, every consequent access to a record protected with
a multi company `ir.rule` fails for that company.

Here, the module automatically creates a project in that company
and later read that project to check a constraint.
The read raises an AccessError, hence the company is not created.

On the other hand, a user without project creation access rights
but erp manager could want to create a company. It should work
in that case too, as the project creation is transparent to the
user and the project is archived.

closes odoo/odoo#34920

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>