[FIX] point_of_sale: fiscal position map to empty tax
This commit fixes the following faulty behaviors:

1. There is a traceback in the pos ui when the selected
customer has fiscal position that maps a tax to nothing.
To prevent this behavior, we avoid to add the undefined
tax destination.

2. When you change the customer in pos ui from someone who
has fiscal position to someone without, the tax in the
order remains to be based on the previous customer. The
tax should change to original tax of the product or
to the tax mapped by the default fiscal position of
the session. To avoid this behavior, we now set the
default fiscal position on the order when a customer has
no fiscal position.

closes odoo/odoo#34955

Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>